Introduction of the "bongcloud" design
Introduction of the "bongcloud" design. GC_PC "bongcloud" drive is a Gustatory + Piriform Cortex drive designed to work with the new Open Ephys hirose EIB and matching amplifier, while minimizing height, as well as being capable of having two drives implanted simultaneously for a total of 128 channel capacity. The bongcloud design is also fully symmetrical, eliminating redundant "left" and "right" designs. The drive is also designed to work with a standard implant tool shared with the GC=GC drive, and adheres to the new 24mm diameter drive basket blech-lab standard.